backend/batch/requirements.txt
feedparser
supabase
python-dotenv
requests
beautifulsoup4
newsapi-python